fix: lists and hashtags pages
