+++ pages/hashtags.vue
@@ -1,14 +1,5 @@
 <script setup lang="ts">
-definePageMeta({
-  middleware: 'auth',
-})
-
 const { t } = useI18n()
-
-const { client } = useMasto()
-const paginator = client.value.v1.followedTags.list({
-  limit: 20,
-})
 
 useHydratedHead({
   title: () => t('nav.hashtags'),
@@ -24,6 +15,6 @@ useHydratedHead({
       </NuxtLink>
     </template>
 
-    <TagCardPaginator v-bind="{ paginator }" />
+    <NuxtPage v-if="isHydrated && currentUser" />
   </MainContent>
 </template>
